Split out path monitoring
